refactor(des): split CheckinQualitativoTab (640→104) em hook + componentes
Extrai a lógica (6 queries DES, init de respostas, derivações de desconto,
salvarCheckin projeção/confirmação) para o hook useCheckinQualitativo e quebra
o JSX em DescontoProjetadoCard, CriterioCard (unifica os cards qualitativo/bônus
que eram idênticos), HistoricoCheckins e ConfirmacaoAndreDialog. Tipos + format
(fmtPct/fmtDate) extraídos. Comportamento 1:1; +13 testes.
