+++ src/components/des/checkinQualitativo/ConfirmacaoAndreDialog.tsx
@@ -1,0 +1,48 @@
+// Dialog de confirmação final com André.
+// Extraído verbatim de src/components/des/CheckinQualitativoTab.tsx (god-component split).
+import { Loader2 } from "lucide-react";
+import {
+  AlertDialog,
+  AlertDialogAction,
+  AlertDialogCancel,
+  AlertDialogContent,
+  AlertDialogDescription,
+  AlertDialogFooter,
+  AlertDialogHeader,
+  AlertDialogTitle,
+} from "@/components/ui/alert-dialog";
+
+interface ConfirmacaoAndreDialogProps {
+  open: boolean;
+  onOpenChange: (open: boolean) => void;
+  saving: boolean;
+  onConfirm: () => void;
+}
+
+export function ConfirmacaoAndreDialog({ open, onOpenChange, saving, onConfirm }: ConfirmacaoAndreDialogProps) {
+  return (
+    <AlertDialog open={open} onOpenChange={onOpenChange}>
+      <AlertDialogContent>
+        <AlertDialogHeader>
+          <AlertDialogTitle>Confirmação final com André</AlertDialogTitle>
+          <AlertDialogDescription>
+            Esta é a avaliação final do trimestre feita com André. Substituirá a projeção atual. Confirmar?
+          </AlertDialogDescription>
+        </AlertDialogHeader>
+        <AlertDialogFooter>
+          <AlertDialogCancel disabled={saving}>Cancelar</AlertDialogCancel>
+          <AlertDialogAction
+            onClick={(e) => {
+              e.preventDefault();
+              onConfirm();
+            }}
+            disabled={saving}
+          >
+            {saving ? <Loader2 className="h-4 w-4 mr-2 animate-spin" /> : null}
+            Confirmar
+          </AlertDialogAction>
+        </AlertDialogFooter>
+      </AlertDialogContent>
+    </AlertDialog>
+  );
+}
